onetimesecret/requirements.txt
Django==3.2.16
python-dotenv==0.21.0
djangorestframework==3.14
django-cryptography==1.1
psycopg2-binary==2.9.5
gunicorn==20.0.4
celery==5.2.7
importlib-metadata==4.13.0
redis==4.4.2